refactor: split out post components
