+++ src/components/PostDetails.tsx
@@ -1,0 +1,52 @@
+import classNames from "classnames";
+import { decode } from "html-entities";
+
+import { useMainContext } from "../MainApp";
+import type { BooruPost } from "../lib/booru";
+
+export const PostDetails = ({ post }: { post: BooruPost }) => {
+	const { query, tempQuery, setTempQuery } = useMainContext();
+
+	const handleTagClick = (tag: string) => {
+		if (!tempQuery.split(" ").includes(tag)) {
+			setTempQuery((tempQuery) =>
+				[...tempQuery.split(" "), tag].filter((e) => e).join(" "),
+			);
+			return;
+		}
+		setTempQuery((tempQuery) =>
+			tempQuery
+				.split(" ")
+				.filter((e) => e && e !== tag)
+				.join(" "),
+		);
+	};
+
+	return (
+		<div className="px-2 break-words text-white col-span-full text-center">
+			{post.tags
+				?.map((e) => decode(e))
+				?.map((tag) => (
+					<button
+						type="button"
+						key={tag}
+						onClick={() => handleTagClick(tag)}
+						className={classNames(
+							"bg-blue-500 hover:bg-blue-700 text-white text-xs font-semibold p-1 px-3 m-1 rounded-full",
+							{
+								"bg-blue-700": query.split(" ").includes(tag),
+								"bg-purple-500":
+									tempQuery.split(" ").includes(tag) &&
+									!query.split(" ").includes(tag),
+								"bg-red-500":
+									!tempQuery.split(" ").includes(tag) &&
+									query.split(" ").includes(tag),
+							},
+						)}
+					>
+						{tag}
+					</button>
+				))}
+		</div>
+	);
+};
